Image_captioning_-_Segmentation-master/requirements.txt
# Core Streamlit
streamlit>=1.30.0

# Deep Learning Frameworks
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0

# Computer Vision
opencv-python-headless>=4.8.0
Pillow>=10.0.0

# COCO Dataset Utils
pycocotools>=2.0.7

# Data Science
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Progress bars
tqdm>=4.65.0

# Model download
gdown>=4.7.1

# NLP
nltk>=3.8.0
